[Rename #114] 모달 이름 변경
