+++ components/mafia/GroupModal.tsx
